graphic added to background
added the page icon graphic to the backgrounds of all the sections on the page. created js to change the color of the menu icon when section is in view. organized js files to own folder. fixed some resizing issues. created new name animation sections for both small and large screens